refactor(portal): extract shares section into SharesSection.svelte
Pull the share-links section (visitor/listener stats) out of the portal
page into a self-loading, co-located component. Parent drops from 2127 to
1598 lines.

The section is fully self-contained: it owns its shares/loadingShares/
sharesTotal/sharesHasMore/loadingMoreShares/expandedShareCode state, the
ShareLinkUser/ShareLinkStats interfaces, and loadShares()/toggleShareDetails(),
and self-loads via onMount(loadShares) like PlaylistsSection. No props,
no callbacks.
