+++ loq.toml
@@ -313,3 +313,7 @@ max_lines = 610
 [[rules]]
 path = "frontend/src/lib/components/portal/TracksSection.svelte"
 max_lines = 1740
+
+[[rules]]
+path = "frontend/src/lib/components/portal/SharesSection.svelte"
+max_lines = 585
